% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincog.R
\name{braincog}
\alias{braincog}
\title{Differential sparse canonical correlation analysis.}
\usage{
braincog(fac, morphometry, cognition, gray_matter, top = 100,
  num_perm = 1000, alpha = 0.05, slurm = FALSE, num_cores = 4,
  seed = 56026)
}
\description{
Differential sparse canonical correlation analysis.
}
