% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size_weighted.R
\name{cluster_size_weighted}
\alias{cluster_size_weighted}
\title{kth largest cluster size measured in absolute values of coefficients from segmented array.}
\usage{
cluster_size_weighted(k, seg_arr, coeff_arr)
}
\description{
kth largest cluster size measured in absolute values of coefficients from segmented array.
}
