% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cca_da2.R
\name{compute_cca_da2}
\alias{compute_cca_da2}
\title{Compute difference vector.}
\usage{
compute_cca_da2(fac, morphometry, cognition, gray_matter, penaltyz = NULL,
  return_seg = FALSE, top = 1000)
}
\description{
Compute difference vector.
}
