% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cognition.R
\name{plot_cognition}
\alias{plot_cognition}
\title{Plot permutation results of cognitive tests.}
\usage{
plot_cognition(fit, domain_mapping, alpha = 0.05)
}
\description{
Plot permutation results of cognitive tests.
}
